+++ src/components/layouts/Mat.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Mat = () => {
+  return (
+    <div>
+      <h2>Mat</h2>
+    </div>
+  );
+};
+
+export default Mat;
