+++ src/components/pages/App.js
@@ -1,23 +1,20 @@
 import { Switch, Route } from 'react-router-dom';
 
-import { Navigation } from '../layouts/Navigation';
-
 import '../../assets/scss/style.scss';
+
+import Home from './Home';
+import UiKit from './UiKit';
 
 function App() {
   return (
     <section className="app">
       <Switch>
         <Route exact path="/">
-          <header className="app-header">
-            <h1>hello world</h1>
-          </header>
-
-          <Navigation />
+          <Home />
         </Route>
 
         <Route exact path="/ui-kit">
-          <h1>ui kit</h1>
+          <UiKit />
         </Route>
       </Switch>
     </section>
