+++ src/components/pages/Home.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import Mat from '../layouts/Mat';
+import Navigation from '../layouts/Navigation';
+
+const Home = () => {
+  return (
+    <>
+      <Mat />
+      <Navigation />
+    </>
+  );
+};
+
+export default Home;
