+++ src/components/pages/UiKit.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const UiKit = () => {
+  return (
+    <div>
+      <h2>ui kit</h2>
+    </div>
+  );
+};
+
+export default UiKit;
