+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 
-import App from './App';
+import App from './components/pages/App';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
